backend/requirements.txt
# ========================================
# MediVault Backend Requirements
# ========================================

# ----------------------------------------
# Google Gemini (Primary AI Engine)
# ----------------------------------------
google-generativeai>=0.3.0

# ----------------------------------------
# Web Framework
# ----------------------------------------
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.2.0

# ----------------------------------------
# Database
# ----------------------------------------
supabase>=2.0.0
psycopg2-binary>=2.9.9

# ----------------------------------------
# Vector Database
# ----------------------------------------
chromadb>=0.4.22

# ----------------------------------------
# Authentication & Security
# ----------------------------------------
PyJWT>=2.8.0
bcrypt>=4.1.0
python-dotenv>=1.0.0

# ----------------------------------------
# File Processing
# ----------------------------------------
Pillow>=10.2.0
PyPDF2>=3.0.0
python-magic>=0.4.27

# ----------------------------------------
# Data Processing
# ----------------------------------------
pandas>=2.1.4
numpy>=1.26.3

# ----------------------------------------
# Utilities
# ----------------------------------------
python-dateutil>=2.8.2
pydantic>=2.5.3
requests>=2.31.0

# ----------------------------------------
# Logging & Monitoring
# ----------------------------------------
loguru>=0.7.2

# ----------------------------------------
# Testing
# ----------------------------------------
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.23.3

# ----------------------------------------
# Development
# ----------------------------------------
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0

# ----------------------------------------
# Optional: OpenAI (Legacy)
# ----------------------------------------
# openai>=1.7.2
